Atualizando para a ultima versão do rasa boilerplate mas começando a alterar as partes que ja tinhamos feito.
