--- Rasa/bot/credentials.yml
@@ -1,4 +1,0 @@
-telegram:
-  access_token: ${TELEGRAM_TOKEN}
-  verify: ${TELEGRAM_BOT_USERNAME}
-  webhook_url: ${TELEGRAM_WEBHOOK}
