--- Rasa/coach/data/intents/actions.md
@@ -1,9 +1,0 @@
-## intent:action_start
-- /start
-- start
-
-## intent:teste
-- testes
-- teste
-- actions
-- action
