--- Rasa/coach/data/stories/actions.md
@@ -1,4 +1,0 @@
-## Start
-* start{"command":"start"}
-    - action_start
-    - utter_start
