--- Rasa/coach/data/stories/chitchat.md
@@ -1,67 +1,0 @@
-## Default
- * default
-    - utter_default
-## Cumprimentar
-* cumprimentar
-    - utter_cumprimentar
-
-## Despedir
-* despedir
-    - utter_despedir
-
-## Tudo_Bem
-* tudo_bem
-    - utter_tudo_bem
-
-## Preocupacao 
-* preocupacao
-    - utter_preocupacao
-
-## Concordar
-* concordar
-    - utter_concordar
-
-
-## Explicacao
-* explicacao
-    - utter_explicacao
-
-## Segredo
-* segredo
-    - utter_segredo
-
-## Risada
-* risada
-    - utter_risada
-
-## Animes1
-* animes1
-    - utter_animes1
-
-## Animes2
-* animes2
-    - utter_animes2
-
-## Xingamento
-* xingamento
-    - utter_xingamento
-
-## Reacao_triste
-* reacao_triste
-    - utter_reacao_triste
-
-## Reacao_feliz
-* reacao_feliz
-    - utter_reacao_feliz
-
-## Series
-* series
-    - utter_series
-
-## Futebol
-* futebol
-    - utter_futebol
-
-## Testes
-* teste
-    - action_test
