--- Rasa/coach/data/stories/commands.md
@@ -1,15 +1,0 @@
-## Help
-* help{"command":"help"}
-    - utter_help
-
-## Atividades
-* atividades{"command":"atividades"}
-    - utter_atividades
-
-## Provas
-* provas{"command":"provas"}
-    - utter_provas
-
-## Trabalhos
-* trabalhos{"command":"trabalhos"}
-    - utter_trabalhos
