--- Rasa/coach/nlu_config.yml
@@ -1,10 +1,0 @@
-language: "pt"
-pipeline:
-- name: "WhitespaceTokenizer"
-- name: "RegexFeaturizer"
-- name: "CountVectorsFeaturizer"
-- name: "CountVectorsFeaturizer"
-  analyzer: "char_wb"
-  min_ngram: 1
-  max_ngram: 4
-- name: "EmbeddingIntentClassifier"
