+++ Rasa/credentials.yml
@@ -1,0 +1,4 @@
+telegram:
+  access_token: ${TELEGRAM_TOKEN}
+  verify: ${TELEGRAM_BOT_USERNAME}
+  webhook_url: ${TELEGRAM_WEBHOOK}
