+++ Rasa/data/intents/actions.md
@@ -1,0 +1,9 @@
+## intent:action_start
+- /start
+- start
+
+## intent:teste
+- testes
+- teste
+- actions
+- action
