+++ Rasa/data/stories/actions.md
@@ -1,0 +1,4 @@
+## Start
+* start{"command":"start"}
+    - action_start
+    - utter_start
