+++ Rasa/data/stories/commands.md
@@ -1,0 +1,15 @@
+## Help
+* help{"command":"help"}
+    - utter_help
+
+## Atividades
+* atividades{"command":"atividades"}
+    - utter_atividades
+
+## Provas
+* provas{"command":"provas"}
+    - utter_provas
+
+## Trabalhos
+* trabalhos{"command":"trabalhos"}
+    - utter_trabalhos
