+++ docker/coach.Dockerfile
@@ -1,11 +1,7 @@
 FROM lappis/botrequirements:boilerplate
 
-COPY ./Rasa/coach /coach
+COPY ./Rasa/ /Rasa
 
-RUN mkdir /src_models
-
-WORKDIR /coach
+WORKDIR /Rasa
 
 RUN make train
-
-RUN find /. | grep -E "(__pycache__|\.pyc|\.pyo$)" | xargs rm -rf
