+++ requirements.txt
@@ -1,0 +1,4 @@
+rasa==1.4.2
+nltk==3.4.5
+flake8==3.7.8
+pymongo==3.9.0
